+++ snippets/usage.rs
@@ -13,11 +13,15 @@ compose_idents!(
     // This function is useful to create identifiers that are unique across multiple macro invocations.
     // `hash(0b11001010010111)` will generate the same value even if called twice in the same macro call,
     // but will be different in different macro calls.
-    MY_UNIQUE_STATIC = [hash(0b11001010010111)]; {
+    MY_UNIQUE_STATIC = [hash(0b11001010010111)];
+    MY_FORMATTED_STR = [FOO, _, BAR]; {
     fn my_fn_1() -> u32 {
         123
     }
 
+    // You can use %alias% syntax to replace aliases with their replacements
+    // in string literals and doc-attributes.
+    #[doc = "This is a docstring for %my_fn_2%"]
     fn my_fn_2() -> u32 {
         321
     }
@@ -27,6 +31,8 @@ compose_idents!(
     static MY_SNAKE_CASE_STATIC: u32 = 42;
     static MY_CAMEL_CASE_STATIC: u32 = 42;
     static MY_UNIQUE_STATIC: u32 = 42;
+    // This is an example of string literal formatting.
+    static MY_FORMATTED_STR: &str = "This is %MY_FORMATTED_STR%";
 });
 
 // It's possible to use arguments of declarative macros as parts of the identifiers.
@@ -63,3 +69,4 @@ assert_eq!(FOO_bar, 42);
 assert_eq!(BAR, 42);
 assert_eq!(snake_case, 42);
 assert_eq!(camelCase, 42);
+assert_eq!(FOO_BAR, "This is FOO_BAR");