+++ tests/compile/format_lit_str.rs
@@ -1,0 +1,11 @@
+use compose_idents::compose_idents;
+
+compose_idents!(
+    my_alias = [foo, _, "baz"]; {
+        static MY_STRING: &str = "Hello, %my_alias%!";
+    };
+);
+
+fn main() {
+    assert_eq!(MY_STRING, "Hello, foo_baz!");
+}